Add compatibility for the mod Lootr, simplify loot table setting.
Using the `setLootTable` member method of a tile entity prevents the mod
Lootr, which replaces loot chests with a chest that creates an inventory
for each player that opens it (making it possible for multiple people to
loot the same chest and all get loot), from being able to hook in.

While a hook can be made into the tile entity, generally the point at
which the `setLootTable` call is made is using the `IWorld` instance of
chunk generation, which isn't the same as the `world` found in the tile
entity (if it even has one).

Conversely, using `LockableLootTileEntity.setLootTable` provides the
chunk generation world and serves as a convenient point for injection.
It otherwise replicates the functionality as you had it: fetching the
tile entity, calling the function and setting a random seed.

Where you also had error testing, I added an additional test that would
trigger your debugging message instead.

Lootr currently supports all vanilla loot chests, including the ones
that are really annoying, and uses a structure processor to convert all
structure NBT files as they're loaded from disk.

Previous attempts to use a tile tracker and determine if a chunk was
ready for external modification (i.e., generation was complete) proved
ineffectual and while some aspects worked well, there were too many
concurrent modification errors in 1.15.

Finally, I'm not sure if I've got the formatting right, but there
doesn't seem to be a huge consistency to the formatting so I tried to
just match what was around.

My only constructive criticism would be that there's a lot of repetitive
functions -- you have two "fillChestBelow" and "fillBarrelBelow"
functions with almost identical code that could just be static methods.