Bugfix: Stale Portal Blocks Send The Player Into The Void
Reported 5 hours ago.  If 1.14 or 1.15 placed a "ReturnPortal" outside of an entrance room, and that dungeon is imported into 1.16 and then encountered, then the portal will pitch the player into the void. This fixes that by making unassigned destinations send the player back to their bed or worldspawn, which was the old functionality anyway.