Room fixes for Theme 2 and other rooms.
Theme 2 gets a barrier roof and a few changes were made to other rooms. Icetrap's floor was fixed, allowing for mobs to spawn behind the doors again. Armor stands in advice_2 and _5 were fixed. And a more obvious exit from the 'treasure trap" in large_slime was added. (The old exit required jumping from a lantern to vines.)