Keyhole Blockstate Cleanup
The old BUILD_STEP is now a boolean IS_BUILDING, since it was only ever being used in that way. The BUILD_PARTICLES property was removed since it duplicated by IS_BUILDING. This also fixes a random crash with the old integer version of the build step. (Probably caused by unloading the keyhole's chunk in the middle of a build.)