New Cheat Structure
All future cheats will be registered under one global "/dimdungeons" cheat. I've seen other mods do this and it is a nice way to keep the command list uncluttered. It is also a self-documenting way of knowing which cheat was added by which mod. (If you have a modpack with 200 mods, and you see /givekey in the command list, you might have no idea what it does or who added it.)